last digit

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -19,20 +19,20 @@ int main(void)
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
 	/* your code goes there */
-	int lastDigit = n % 10;
+	double lastDigit = n % 10;
 
 	if (lastDigit > 5)
 	{
-		printf("Last digit of %d is %d and is greater than 5", n, lastDigit);
+		printf("Last digit of %d is %lf and is greater than 5", n, lastDigit);
 	}
 	else if (lastDigit < 6 && lastDigit != 0)
 	{
-		printf("Last digit of %d is %d and is less than 6 and not 0\n",
+		printf("Last digit of %d is %lf and is less than 6 and not 0\n",
 			n, lastDigit);
 	}
 	else
 	{
-		printf("Last digit of %d is %d and is 0\n", n, lastDigit);
+		printf("Last digit of %d is %lf and is 0\n", n, lastDigit);
 	}
